pyproject.toml
[tool.poetry]
name = "riseup-vpn-configurator"
version = "1.3.0"
description = "a Bitmask-compatible OpenVPN configurator for Riseup and other known providers"
authors = ["kmille <github@androidloves.me>"]
readme = "README.md"
packages = [{include = "riseup_vpn_configurator"}]
license = "GPL-3.0-only"
repository = "https://github.com/mkiskis/vpn-configurator"
homepage = "https://github.com/mkiskis/vpn-configurator"


[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.2"
requests = "^2.28.2"
pyyaml = "^6.0"
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.1"
psutil = "^6.0.0"
icmplib = "^3.0.4"


[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
pytest = "^8.3.2"
coverage = "^7.1.0"
flake8 = "^7.1.1"


[tool.poetry.group.mypy.dependencies]
mypy = "^1.0.0"
types-pyyaml = "^6.0.12.6"
types-requests = "^2.28.11.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
riseup-vpn-configurator = 'riseup_vpn_configurator.__init__:main'

[tool.setuptools.package-data]
riseup_vpn_configurator = ["riseup-vpn.yaml"]

[tool.mypy]
ignore_missing_imports = true
